+++ crates/core-utils/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "embucket_utils"
+name = "core-utils"
 version = "0.1.0"
 edition = "2021"
 license-file = { workspace = true }
