+++ crates/core-utils/src/iterable.rs
