+++ crates/core-utils/src/scan_iterator.rs
