+++ crates/datafusion-builtins/src/aggregate/booland_agg.rs
@@ -1,5 +1,5 @@
-use crate::execution::datafusion::functions::aggregate::macros::make_udaf_function;
-use crate::execution::datafusion::functions::array_to_boolean;
+use crate::aggregate::macros::make_udaf_function;
+use crate::array_to_boolean;
 use datafusion::arrow::array::ArrayRef;
 use datafusion::arrow::datatypes::DataType;
 use datafusion::error::Result as DFResult;
@@ -9,25 +9,24 @@ use datafusion_expr::function::AccumulatorArgs;
 use datafusion_expr::{AggregateUDFImpl, Signature, Volatility};
 use std::any::Any;
 
-/// Boolxor function
-/// Returns TRUE if exactly one Boolean record in the group evaluates to TRUE.
-///
+/// Booland Agg function
+/// Returns TRUE if all non-NULL Boolean records in a group evaluate to TRUE.
 /// If all records in the group are NULL, or if the group is empty, the function returns NULL.
 ///
-/// Syntax: `boolxor_agg(<expr>)`
+/// Syntax: `booland_agg(<expr>)`
 
 #[derive(Debug, Clone)]
-pub struct BoolXorAggUDAF {
+pub struct BoolAndAggUDAF {
     signature: Signature,
 }
 
-impl Default for BoolXorAggUDAF {
+impl Default for BoolAndAggUDAF {
     fn default() -> Self {
         Self::new()
     }
 }
 
-impl BoolXorAggUDAF {
+impl BoolAndAggUDAF {
     pub fn new() -> Self {
         Self {
             signature: Signature::any(1, Volatility::Immutable),
@@ -35,13 +34,13 @@ impl BoolXorAggUDAF {
     }
 }
 
-impl AggregateUDFImpl for BoolXorAggUDAF {
+impl AggregateUDFImpl for BoolAndAggUDAF {
     fn as_any(&self) -> &dyn Any {
         self
     }
 
     fn name(&self) -> &'static str {
-        "boolxor_agg"
+        "booland_agg"
     }
 
     fn signature(&self) -> &Signature {
@@ -53,22 +52,22 @@ impl AggregateUDFImpl for BoolXorAggUDAF {
     }
 
     fn accumulator(&self, _acc_args: AccumulatorArgs) -> DFResult<Box<dyn Accumulator>> {
-        Ok(Box::new(BoolXorAggAccumulator::new()))
+        Ok(Box::new(BoolAndAggAccumulator::new()))
     }
 }
 
 #[derive(Debug)]
-struct BoolXorAggAccumulator {
+struct BoolAndAggAccumulator {
     state: Option<bool>,
 }
 
-impl BoolXorAggAccumulator {
+impl BoolAndAggAccumulator {
     pub const fn new() -> Self {
         Self { state: None }
     }
 }
 
-impl Accumulator for BoolXorAggAccumulator {
+impl Accumulator for BoolAndAggAccumulator {
     fn update_batch(&mut self, values: &[ArrayRef]) -> DFResult<()> {
         if values.is_empty() {
             return Ok(());
@@ -78,21 +77,18 @@ impl Accumulator for BoolXorAggAccumulator {
         }
 
         let barr = array_to_boolean(&values[0])?;
-        let mut is_null = true;
+        let mut non_null = false;
         for val in &barr {
             if val.is_some() {
-                is_null = false;
+                non_null = true;
             }
-            if matches!(val, Some(true)) {
-                if matches!(self.state, Some(true)) {
-                    self.state = Some(false);
-                    return Ok(());
-                }
-                self.state = Some(true);
+            if matches!(val, Some(false)) {
+                self.state = Some(false);
+                return Ok(());
             }
         }
-        if !is_null && !matches!(self.state, Some(true)) {
-            self.state = Some(false);
+        if non_null {
+            self.state = Some(true);
         }
 
         Ok(())
@@ -115,31 +111,27 @@ impl Accumulator for BoolXorAggAccumulator {
             return Ok(());
         }
 
-        let mut is_null = true;
+        let mut non_null = false;
         for state in states {
             let v = ScalarValue::try_from_array(state, 0)?;
             if !v.is_null() {
-                is_null = false;
+                non_null = true;
             }
-            if matches!(v, ScalarValue::Boolean(Some(true))) {
-                if matches!(self.state, Some(true)) {
-                    self.state = Some(false);
-                    return Ok(());
-                }
-
-                self.state = Some(true);
+            if matches!(v, ScalarValue::Boolean(Some(false))) {
+                self.state = Some(false);
+                return Ok(());
             }
         }
 
-        if !is_null && !matches!(self.state, Some(true)) {
-            self.state = Some(false);
+        if non_null {
+            self.state = Some(true);
         }
 
         Ok(())
     }
 }
 
-make_udaf_function!(BoolXorAggUDAF);
+make_udaf_function!(BoolAndAggUDAF);
 
 #[cfg(test)]
 mod tests {
@@ -152,42 +144,42 @@ mod tests {
 
     #[tokio::test]
     async fn test_merge() -> DFResult<()> {
-        let mut acc = BoolXorAggAccumulator::new();
+        let mut acc = BoolAndAggAccumulator::new();
         acc.merge_batch(&[
             Arc::new(BooleanArray::from(vec![Some(true)])),
             Arc::new(BooleanArray::from(vec![Some(true)])),
-        ])?;
-        assert_eq!(acc.state, Some(false));
-
-        let mut acc = BoolXorAggAccumulator::new();
-        acc.merge_batch(&[
-            Arc::new(BooleanArray::from(vec![Some(true)])),
-            Arc::new(BooleanArray::from(vec![Some(false)])),
         ])?;
         assert_eq!(acc.state, Some(true));
 
-        let mut acc = BoolXorAggAccumulator::new();
+        let mut acc = BoolAndAggAccumulator::new();
+        acc.merge_batch(&[
+            Arc::new(BooleanArray::from(vec![Some(true)])),
+            Arc::new(BooleanArray::from(vec![Some(false)])),
+        ])?;
+        assert_eq!(acc.state, Some(false));
+
+        let mut acc = BoolAndAggAccumulator::new();
         acc.merge_batch(&[
             Arc::new(BooleanArray::from(vec![Some(false)])),
             Arc::new(BooleanArray::from(vec![Some(false)])),
         ])?;
         assert_eq!(acc.state, Some(false));
 
-        let mut acc = BoolXorAggAccumulator::new();
+        let mut acc = BoolAndAggAccumulator::new();
         acc.merge_batch(&[
             Arc::new(BooleanArray::from(vec![Some(true)])),
             Arc::new(BooleanArray::from(vec![None])),
         ])?;
         assert_eq!(acc.state, Some(true));
 
-        let mut acc = BoolXorAggAccumulator::new();
+        let mut acc = BoolAndAggAccumulator::new();
         acc.merge_batch(&[
             Arc::new(BooleanArray::from(vec![Some(false)])),
             Arc::new(BooleanArray::from(vec![None])),
         ])?;
         assert_eq!(acc.state, Some(false));
 
-        let mut acc = BoolXorAggAccumulator::new();
+        let mut acc = BoolAndAggAccumulator::new();
         acc.merge_batch(&[
             Arc::new(BooleanArray::from(vec![None])),
             Arc::new(BooleanArray::from(vec![None])),
@@ -201,7 +193,7 @@ mod tests {
     async fn test_sql() -> DFResult<()> {
         let config = SessionConfig::new();
         let ctx = SessionContext::new_with_config(config);
-        ctx.register_udaf(AggregateUDF::from(BoolXorAggUDAF::new()));
+        ctx.register_udaf(AggregateUDF::from(BoolAndAggUDAF::new()));
         ctx.sql(
             "create table test_boolean_agg
 (
@@ -222,7 +214,7 @@ mod tests {
         .await?;
 
         let result = ctx
-            .sql("select id, boolxor_agg(c) from test_boolean_agg group by id order by id;")
+            .sql("select id, booland_agg(c) from test_boolean_agg group by id order by id;")
             .await?
             .collect()
             .await?;
@@ -230,10 +222,58 @@ mod tests {
         assert_batches_eq!(
             &[
                 "+----+---------------------------------+",
-                "| id | boolxor_agg(test_boolean_agg.c) |",
+                "| id | booland_agg(test_boolean_agg.c) |",
                 "+----+---------------------------------+",
-                "| 1  | false                           |",
-                "| 2  | true                            |",
+                "| 1  | true                            |",
+                "| 2  | false                           |",
+                "| 3  | true                            |",
+                "| 4  | false                           |",
+                "| 5  |                                 |",
+                "+----+---------------------------------+",
+            ],
+            &result
+        );
+
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn test_numeric() -> DFResult<()> {
+        let config = SessionConfig::new();
+        let ctx = SessionContext::new_with_config(config);
+        ctx.register_udaf(AggregateUDF::from(BoolAndAggUDAF::new()));
+        ctx.sql(
+            "create table test_boolean_agg
+(
+    id integer,
+    c  integer
+) as values
+    (1, 1),
+    (1, 1),
+    (2, 1),
+    (2, 0),
+    (3, 1),
+    (3, null),
+    (4, 0),
+    (4, null),
+    (5, null),
+    (5, null);",
+        )
+        .await?;
+
+        let result = ctx
+            .sql("select id, booland_agg(c) from test_boolean_agg group by id order by id;")
+            .await?
+            .collect()
+            .await?;
+
+        assert_batches_eq!(
+            &[
+                "+----+---------------------------------+",
+                "| id | booland_agg(test_boolean_agg.c) |",
+                "+----+---------------------------------+",
+                "| 1  | true                            |",
+                "| 2  | false                           |",
                 "| 3  | true                            |",
                 "| 4  | false                           |",
                 "| 5  |                                 |",
