+++ crates/datafusion-builtins/src/aggregate/boolxor_agg.rs
@@ -1,5 +1,5 @@
-use crate::execution::datafusion::functions::aggregate::macros::make_udaf_function;
-use crate::execution::datafusion::functions::array_to_boolean;
+use crate::aggregate::macros::make_udaf_function;
+use crate::array_to_boolean;
 use datafusion::arrow::array::ArrayRef;
 use datafusion::arrow::datatypes::DataType;
 use datafusion::error::Result as DFResult;
@@ -9,25 +9,25 @@ use datafusion_expr::function::AccumulatorArgs;
 use datafusion_expr::{AggregateUDFImpl, Signature, Volatility};
 use std::any::Any;
 
-/// Boolor function
-/// Returns TRUE if at least one Boolean record in a group evaluates to TRUE.
+/// Boolxor function
+/// Returns TRUE if exactly one Boolean record in the group evaluates to TRUE.
 ///
 /// If all records in the group are NULL, or if the group is empty, the function returns NULL.
 ///
-/// Syntax: `boolor_agg(<expr>)`
+/// Syntax: `boolxor_agg(<expr>)`
 
 #[derive(Debug, Clone)]
-pub struct BoolOrAggUDAF {
+pub struct BoolXorAggUDAF {
     signature: Signature,
 }
 
-impl Default for BoolOrAggUDAF {
+impl Default for BoolXorAggUDAF {
     fn default() -> Self {
         Self::new()
     }
 }
 
-impl BoolOrAggUDAF {
+impl BoolXorAggUDAF {
     pub fn new() -> Self {
         Self {
             signature: Signature::any(1, Volatility::Immutable),
@@ -35,13 +35,13 @@ impl BoolOrAggUDAF {
     }
 }
 
-impl AggregateUDFImpl for BoolOrAggUDAF {
+impl AggregateUDFImpl for BoolXorAggUDAF {
     fn as_any(&self) -> &dyn Any {
         self
     }
 
     fn name(&self) -> &'static str {
-        "boolor_agg"
+        "boolxor_agg"
     }
 
     fn signature(&self) -> &Signature {
@@ -53,22 +53,22 @@ impl AggregateUDFImpl for BoolOrAggUDAF {
     }
 
     fn accumulator(&self, _acc_args: AccumulatorArgs) -> DFResult<Box<dyn Accumulator>> {
-        Ok(Box::new(BoolOrAggAccumulator::new()))
+        Ok(Box::new(BoolXorAggAccumulator::new()))
     }
 }
 
 #[derive(Debug)]
-struct BoolOrAggAccumulator {
+struct BoolXorAggAccumulator {
     state: Option<bool>,
 }
 
-impl BoolOrAggAccumulator {
+impl BoolXorAggAccumulator {
     pub const fn new() -> Self {
         Self { state: None }
     }
 }
 
-impl Accumulator for BoolOrAggAccumulator {
+impl Accumulator for BoolXorAggAccumulator {
     fn update_batch(&mut self, values: &[ArrayRef]) -> DFResult<()> {
         if values.is_empty() {
             return Ok(());
@@ -78,18 +78,20 @@ impl Accumulator for BoolOrAggAccumulator {
         }
 
         let barr = array_to_boolean(&values[0])?;
-
-        let mut non_null = false;
+        let mut is_null = true;
         for val in &barr {
             if val.is_some() {
-                non_null = true;
+                is_null = false;
             }
             if matches!(val, Some(true)) {
+                if matches!(self.state, Some(true)) {
+                    self.state = Some(false);
+                    return Ok(());
+                }
                 self.state = Some(true);
-                return Ok(());
             }
         }
-        if non_null {
+        if !is_null && !matches!(self.state, Some(true)) {
             self.state = Some(false);
         }
 
@@ -113,19 +115,23 @@ impl Accumulator for BoolOrAggAccumulator {
             return Ok(());
         }
 
-        let mut non_null = false;
+        let mut is_null = true;
         for state in states {
             let v = ScalarValue::try_from_array(state, 0)?;
             if !v.is_null() {
-                non_null = true;
+                is_null = false;
             }
             if matches!(v, ScalarValue::Boolean(Some(true))) {
+                if matches!(self.state, Some(true)) {
+                    self.state = Some(false);
+                    return Ok(());
+                }
+
                 self.state = Some(true);
-                return Ok(());
             }
         }
 
-        if non_null {
+        if !is_null && !matches!(self.state, Some(true)) {
             self.state = Some(false);
         }
 
@@ -133,7 +139,7 @@ impl Accumulator for BoolOrAggAccumulator {
     }
 }
 
-make_udaf_function!(BoolOrAggUDAF);
+make_udaf_function!(BoolXorAggUDAF);
 
 #[cfg(test)]
 mod tests {
@@ -146,42 +152,42 @@ mod tests {
 
     #[tokio::test]
     async fn test_merge() -> DFResult<()> {
-        let mut acc = BoolOrAggAccumulator::new();
+        let mut acc = BoolXorAggAccumulator::new();
         acc.merge_batch(&[
             Arc::new(BooleanArray::from(vec![Some(true)])),
             Arc::new(BooleanArray::from(vec![Some(true)])),
         ])?;
-        assert_eq!(acc.state, Some(true));
+        assert_eq!(acc.state, Some(false));
 
-        let mut acc = BoolOrAggAccumulator::new();
+        let mut acc = BoolXorAggAccumulator::new();
         acc.merge_batch(&[
             Arc::new(BooleanArray::from(vec![Some(true)])),
             Arc::new(BooleanArray::from(vec![Some(false)])),
         ])?;
         assert_eq!(acc.state, Some(true));
 
-        let mut acc = BoolOrAggAccumulator::new();
+        let mut acc = BoolXorAggAccumulator::new();
         acc.merge_batch(&[
             Arc::new(BooleanArray::from(vec![Some(false)])),
             Arc::new(BooleanArray::from(vec![Some(false)])),
         ])?;
         assert_eq!(acc.state, Some(false));
 
-        let mut acc = BoolOrAggAccumulator::new();
+        let mut acc = BoolXorAggAccumulator::new();
         acc.merge_batch(&[
             Arc::new(BooleanArray::from(vec![Some(true)])),
             Arc::new(BooleanArray::from(vec![None])),
         ])?;
         assert_eq!(acc.state, Some(true));
 
-        let mut acc = BoolOrAggAccumulator::new();
+        let mut acc = BoolXorAggAccumulator::new();
         acc.merge_batch(&[
             Arc::new(BooleanArray::from(vec![Some(false)])),
             Arc::new(BooleanArray::from(vec![None])),
         ])?;
         assert_eq!(acc.state, Some(false));
 
-        let mut acc = BoolOrAggAccumulator::new();
+        let mut acc = BoolXorAggAccumulator::new();
         acc.merge_batch(&[
             Arc::new(BooleanArray::from(vec![None])),
             Arc::new(BooleanArray::from(vec![None])),
@@ -195,7 +201,7 @@ mod tests {
     async fn test_sql() -> DFResult<()> {
         let config = SessionConfig::new();
         let ctx = SessionContext::new_with_config(config);
-        ctx.register_udaf(AggregateUDF::from(BoolOrAggUDAF::new()));
+        ctx.register_udaf(AggregateUDF::from(BoolXorAggUDAF::new()));
         ctx.sql(
             "create table test_boolean_agg
 (
@@ -216,22 +222,22 @@ mod tests {
         .await?;
 
         let result = ctx
-            .sql("select id, boolor_agg(c) from test_boolean_agg group by id order by id;")
+            .sql("select id, boolxor_agg(c) from test_boolean_agg group by id order by id;")
             .await?
             .collect()
             .await?;
 
         assert_batches_eq!(
             &[
-                "+----+--------------------------------+",
-                "| id | boolor_agg(test_boolean_agg.c) |",
-                "+----+--------------------------------+",
-                "| 1  | true                           |",
-                "| 2  | true                           |",
-                "| 3  | true                           |",
-                "| 4  | false                          |",
-                "| 5  |                                |",
-                "+----+--------------------------------+",
+                "+----+---------------------------------+",
+                "| id | boolxor_agg(test_boolean_agg.c) |",
+                "+----+---------------------------------+",
+                "| 1  | false                           |",
+                "| 2  | true                            |",
+                "| 3  | true                            |",
+                "| 4  | false                           |",
+                "| 5  |                                 |",
+                "+----+---------------------------------+",
             ],
             &result
         );
