+++ crates/datafusion-builtins/src/booland.rs
@@ -1,4 +1,4 @@
-use crate::execution::datafusion::functions::array_to_boolean;
+use crate::array_to_boolean;
 use datafusion::arrow::array::builder::BooleanBuilder;
 use datafusion::arrow::datatypes::DataType;
 use datafusion::error::Result as DFResult;
@@ -102,15 +102,15 @@ mod tests {
         let result = ctx.sql(q).await?.collect().await?;
 
         assert_batches_eq!(
-        &[
-"+-----------------------------+---------------------------------+----------------------------+------------------------+------------------------+--------------------+",
-"| booland(Int64(1),Int64(-2)) | booland(Int64(0),Float64(2.35)) | booland(Int64(0),Int64(0)) | booland(Int64(0),NULL) | booland(NULL,Int64(3)) | booland(NULL,NULL) |",
-"+-----------------------------+---------------------------------+----------------------------+------------------------+------------------------+--------------------+",
-"| true                        | false                           | false                      | false                  |                        |                    |",
-"+-----------------------------+---------------------------------+----------------------------+------------------------+------------------------+--------------------+",
-        ],
-        &result
-    );
+            &[
+                "+-----------------------------+---------------------------------+----------------------------+------------------------+------------------------+--------------------+",
+                "| booland(Int64(1),Int64(-2)) | booland(Int64(0),Float64(2.35)) | booland(Int64(0),Int64(0)) | booland(Int64(0),NULL) | booland(NULL,Int64(3)) | booland(NULL,NULL) |",
+                "+-----------------------------+---------------------------------+----------------------------+------------------------+------------------------+--------------------+",
+                "| true                        | false                           | false                      | false                  |                        |                    |",
+                "+-----------------------------+---------------------------------+----------------------------+------------------------+------------------------+--------------------+",
+            ],
+            &result
+        );
 
         Ok(())
     }
