+++ crates/datafusion-builtins/src/boolor.rs
@@ -1,5 +1,5 @@
-use crate::execution::datafusion::functions::array_to_boolean;
-use crate::execution::datafusion::functions::booland::is_true;
+use crate::array_to_boolean;
+use crate::booland::is_true;
 use datafusion::arrow::array::builder::BooleanBuilder;
 use datafusion::arrow::datatypes::DataType;
 use datafusion::error::Result as DFResult;
@@ -99,15 +99,15 @@ mod tests {
         let result = ctx.sql(q).await?.collect().await?;
 
         assert_batches_eq!(
-        &[
-"+---------------------------+---------------------------------+-----------------------+---------------------------+-----------------------+-------------------+",
-"| boolor(Int64(1),Int64(2)) | boolor(Float64(-1.35),Int64(0)) | boolor(Int64(3),NULL) | boolor(Int64(0),Int64(0)) | boolor(NULL,Int64(0)) | boolor(NULL,NULL) |",
-"+---------------------------+---------------------------------+-----------------------+---------------------------+-----------------------+-------------------+",
-"| true                      | true                            | true                  | false                     |                       |                   |",
-"+---------------------------+---------------------------------+-----------------------+---------------------------+-----------------------+-------------------+",
-        ],
-        &result
-    );
+            &[
+                "+---------------------------+---------------------------------+-----------------------+---------------------------+-----------------------+-------------------+",
+                "| boolor(Int64(1),Int64(2)) | boolor(Float64(-1.35),Int64(0)) | boolor(Int64(3),NULL) | boolor(Int64(0),Int64(0)) | boolor(NULL,Int64(0)) | boolor(NULL,NULL) |",
+                "+---------------------------+---------------------------------+-----------------------+---------------------------+-----------------------+-------------------+",
+                "| true                      | true                            | true                  | false                     |                       |                   |",
+                "+---------------------------+---------------------------------+-----------------------+---------------------------+-----------------------+-------------------+",
+            ],
+            &result
+        );
         Ok(())
     }
 }
