+++ crates/datafusion-builtins/src/boolxor.rs
@@ -1,5 +1,5 @@
-use crate::execution::datafusion::functions::array_to_boolean;
-use crate::execution::datafusion::functions::booland::is_true;
+use crate::array_to_boolean;
+use crate::booland::is_true;
 use datafusion::arrow::array::builder::BooleanBuilder;
 use datafusion::arrow::datatypes::DataType;
 use datafusion::error::Result as DFResult;
@@ -106,15 +106,15 @@ mod tests {
         let result = ctx.sql(q).await?.collect().await?;
 
         assert_batches_eq!(
-        &[
-"+----------------------------+-----------------------------+----------------------------+------------------------+------------------------+--------------------+",
-"| boolxor(Int64(2),Int64(0)) | boolxor(Int64(1),Int64(-1)) | boolxor(Int64(0),Int64(0)) | boolxor(NULL,Int64(3)) | boolxor(NULL,Int64(0)) | boolxor(NULL,NULL) |",
-"+----------------------------+-----------------------------+----------------------------+------------------------+------------------------+--------------------+",
-"| true                       | false                       | false                      |                        |                        |                    |",
-"+----------------------------+-----------------------------+----------------------------+------------------------+------------------------+--------------------+",
-        ],
-        &result
-    );
+            &[
+                "+----------------------------+-----------------------------+----------------------------+------------------------+------------------------+--------------------+",
+                "| boolxor(Int64(2),Int64(0)) | boolxor(Int64(1),Int64(-1)) | boolxor(Int64(0),Int64(0)) | boolxor(NULL,Int64(3)) | boolxor(NULL,Int64(0)) | boolxor(NULL,NULL) |",
+                "+----------------------------+-----------------------------+----------------------------+------------------------+------------------------+--------------------+",
+                "| true                       | false                       | false                      |                        |                        |                    |",
+                "+----------------------------+-----------------------------+----------------------------+------------------------+------------------------+--------------------+",
+            ],
+            &result
+        );
         Ok(())
     }
 }
