+++ crates/datafusion-builtins/src/geospatial/accessors/geometry.rs
