+++ crates/datafusion-builtins/src/geospatial/error.rs
