+++ crates/datafusion-builtins/src/table/flatten.rs
@@ -1,20 +1,3 @@
-// Licensed to the Apache Software Foundation (ASF) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The ASF licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
 use datafusion::arrow::array::builder::{StringBuilder, UInt64Builder};
 use datafusion::arrow::array::{ArrayRef, RecordBatch, StringArray, UInt64Array};
 use datafusion::arrow::datatypes::{DataType, Field, Schema, SchemaRef};
@@ -22,14 +5,14 @@ use datafusion::catalog::{TableFunctionImpl, TableProvider};
 use datafusion::datasource::MemTable;
 use datafusion::physical_expr::create_physical_expr;
 use datafusion::physical_plan::ColumnarValue;
-use datafusion_common::{exec_err, DFSchema, DataFusionError, Result as DFResult, ScalarValue};
-use datafusion_expr::execution_props::ExecutionProps;
+use datafusion_common::{DFSchema, DataFusionError, Result as DFResult, ScalarValue, exec_err};
 use datafusion_expr::Expr;
+use datafusion_expr::execution_props::ExecutionProps;
 use serde_json::Value;
 use std::cell::RefCell;
 use std::rc::Rc;
-use std::sync::atomic::{AtomicU64, Ordering};
 use std::sync::Arc;
+use std::sync::atomic::{AtomicU64, Ordering};
 
 #[derive(Debug)]
 enum Mode {
@@ -563,7 +546,7 @@ fn get_args(args: &[&Expr]) -> DFResult<Args> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::execution::datafusion::functions::parse_json::ParseJsonFunc;
+    use crate::parse_json::ParseJsonFunc;
     use datafusion::prelude::SessionContext;
     use datafusion_common::assert_batches_eq;
     use std::sync::Arc;
