--- crates/datafusion-functions/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "datafusion-functions"
-version = "0.1.0"
-edition = "2024"
-license-file.workspace = true
-
-[dependencies]
-
-[lints]
-workspace = true
